requirements.txt
json5
openai
retrying
pyloudnorm
nltk
flask
voicefixer
omegaconf==2.0.6
fairseq==0.12.2
audiolm-pytorch==1.1.4